requirements.txt
streamlit
google-genai
time
